+++ Matrix.cs
@@ -9,42 +9,42 @@
         public int Rows { get { return _rows; } set { _rows = value; } }
         private int _columns;
         public int Columns { get { return _columns; } set { _columns = value; } }
-        public double[,] values;
+        public float[,] values;
 
 
         public Matrix(int size)
         {
             _rows = size;
             _columns = size;
-            values = new double[size, size];    // квадратная
+            values = new float[size, size];    // квадратная
         }
 
         public Matrix(int rows, int columns)
         {
             _rows = rows;
             _columns = columns;
-            values = new double[_rows, _columns];    // строки, столбцы
+            values = new float[_rows, _columns];    // строки, столбцы
         }
 
-        public Matrix(double[] vector)
+        public Matrix(float[] vector)
         {
             _rows = 1;
             _columns = vector.Length;
-            values = new double[1, _columns];
+            values = new float[1, _columns];
             for (int i = 0; i < _columns; i++)
                 values[0, i] = vector[i];
         }
 
-        public Matrix(double[,] vector)
+        public Matrix(float[,] vector)
         {
             _rows = vector.GetUpperBound(0) + 1;
             _columns = vector.GetUpperBound(1) + 1;
-            //values = new double[_rows, _columns];
-            values = (double[,])vector.Clone();
+            //values = new float[_rows, _columns];
+            values = (float[,])vector.Clone();
         }
 
 
-        public double this[int index1, int index2]
+        public float this[int index1, int index2]
         {
             get { return values[index1, index2]; }
             set { values[index1, index2] = value; }
